Reworked examples, bug fixes, improved visualization, documentation update and small improvements. (#135)
* Renamed UniformBoussinesq to NeutralStratification and StratifiedBoussinesq to StableStratification.

* Replaced exit() calls with error(...) calls and included more details in the error messages given by check_rays.

* Fixed a bug in Predictands that caused the initialization of Boussinesq simulations with non-zero density fluctuations to be incorrect.

* Added a Boussinesq method of reset_predictands! that does not reset the density.

* Fixed the initialization of the mass-weighted potential temperature in compressible mode. Removed the atmosphere-namelist parameter initial_thetap.

* Added xlims! and ylims! calls to plot_contours.

* Added a vortex example.

* Follow-up correction.

* Added a cold-bubble example.

* Added a hot-bubble example.

* Changed the default values of x_size, y_size and z_size to 1.

* Unified the vertical spacing between namelist definitions in the example scripts.

* Minor adjustment in docs/make.jl.

* Fixed a bug in Atmosphere (missing pointer).

* Implemented configurable space and time units in plot_contours.

* Updated the docstring of plot_contours.

* Improved plot_contours: Multiple variables, automatic colorbar labels and better formatting of colorbar-tick labels.

* Reformatted the code.

* Added a wave-packet example.

* Improved the labels in plot_contours.

* Reduced the domain size in the cold-bubble example.

* Reduced the domain size in the hot-bubble example.

* Reduced the domain size and velocity in the vortex example.

* Made the periodic-hill example Boussinesq and more nonlinear.

* Set the Coriolis frequency of the wave-packet example to 0.0001.

* Changed the background color of plots in the x-z or y-z plane to black.

* Renamed plot_contours to plot_output and added ray-volume visualization. Changed the run and visualization scripts for the WKB-mountain-wave example accordingly.

* Follow-up bug fix.

* Made plot_output use the labels stored in the output file.

* Replaced the label strings in create_output with LaTeX strings (HDF5 will internally convert them to strings but writing LaTeX strings is much more convenient).

* Reformatted the code.

* Some simplifications in plot_output.

* The plot_output function can now be used to visualize any ray-volume property. It no longer visualizes the physical-space wave-action density.

* Implemented removal of existing output files in create_output (this seems to be necessary in very specific situations).

* Corrected a typo in create_output.

* Moved the wave-packet helper functions to a separate file. Fixed the initialization of the density fluctuations in the wave-packet example. Modified the wave-packet configuration and made in work in parallel.

* Moved the visualization to the run scripts and renamed the submit folder "scripts".

* Added a WKB-wave-packet example.

* Updated the documentation.

* Fixed horizontal spacing in LaTeX equations (previously, it wasn't displayed correctly in GitHub Markdown preview).

* Improved regular expressions for Julia-code-block detection.

* Fixed a bug in check_rays.

+++ docs/make.jl
@@ -8,29 +8,27 @@ using Revise
 using PinCFlow
 
 # Insert the example scripts.
-@ivy for folder in ("examples/submit/", "examples/visualization/"),
-    script_file in readdir(folder)
-
+folder = "examples/scripts/"
+@ivy for script_file in readdir(folder)
     if endswith(script_file, ".jl")
         script = read(folder * script_file, String)
         code = Regex(
-            "(?s)(?<=\\n`{3}julia\\n)# " *
+            "(?sm)(?<=^```julia\\n)# " *
             folder *
-            script_file[1:(end - 3)] *
-            "\\.jl\\n(.(?!\\n`{3}))*.",
+            script_file *
+            "(.(?!^```\\n))*",
         )
         if script_file == "periodic_hill.jl"
             page_file = "README.md"
         else
-            page_file =
-                "docs/src/examples/" *
-                script_file[1:(end - 3)] *
-                "_simulation.md"
+            page_file = "docs/src/examples.md"
         end
-        page = replace(read(page_file, String), code => script)
-        open(page_file, "w") do io
-            write(io, page)
-            return
+        if isfile(page_file)
+            page = replace(read(page_file, String), code => script)
+            open(page_file, "w") do io
+                write(io, page)
+                return
+            end
         end
     end
 end
@@ -143,10 +141,7 @@ makedocs(;
     remotes = nothing,
     pages = [
         "Home" => "index.md",
-        "Examples" => [
-            "Mountain-wave simulation" => "examples/mountain_wave_simulation.md",
-            "WKB mountain-wave simulation" => "examples/wkb_mountain_wave_simulation.md",
-        ],
+        "Examples" => "examples.md",
         "Theory" => [
             "Physics" => "theory/physics.md",
             "Numerics" => "theory/numerics.md",
